V0.1 - Streaking Meteors
Meteors have a chance to spawn high in the sky around one player currently online.

Meteors will streak across the sky in a random direction until they either blow up(after exceeding their randomly-set lifespan) or fizzle out(despawn).

When a meteor explodes, it will make a distinguishable explosion sound and ember particles will burst off of it. It will also drop a small amount of materials that it was made up of, in the form of stones and nuggets.