V0.2 - Meteorites
Now handles the ability for meteors to strike the ground, causing a a crater to form and for the meteor to leave parts of its material in the ground. Will cause terrain blocks and meteor drops to blast back as shrapnel.

Meteors are now forced to be made of metals that exist as in the proper stone types and must have a nugget of its type.

Added config options for the mod:

Destructive,
MinimumMinutesBetweenMeteorSpawns,
MaximumMinutesBetweenMeteorSpawns